requirements.txt
appdirs==1.4.2
Django==1.10.5
djangorestframework==3.6.2
packaging==16.8
pycryptodomex==3.4.5
six==1.10.0
urllib3==1.20
